backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv>=1.0.0
pillow>=11.0.0
httpx>=0.28.0
python-multipart>=0.0.18
mediapipe>=0.10.18
